+++ workspaces/rbac/plugins/rbac-backend/__fixtures__/test-utils.ts
@@ -27,22 +27,21 @@ import {
 import * as Knex from 'knex';
 import { MockClient } from 'knex-mock-client';
 
-import { CasbinDBAdapterFactory } from '../src/database/casbin-adapter-factory';
+import { CasbinKnexAdapter } from '../src/database/casbin-knex-adapter';
 import { RoleMetadataStorage } from '../src/database/role-metadata';
 import { RBACPermissionPolicy } from '../src/policies/permission-policy';
 import { BackstageRoleManager } from '../src/role-manager/role-manager';
 import { DefaultPermissionsReader } from '../src/default-permissions/default-permissions';
 import { EnforcerDelegate } from '../src/service/enforcer-delegate';
 import { MODEL } from '../src/service/permission-model';
-import { PluginPermissionMetadataCollector } from '../src/service/plugin-endpoints';
 import {
   mockAuditorService,
   conditionalStorageMock,
+  createTestCasbinKnex,
   csvPermFile,
   mockAuthService,
   mockUserInfoService,
   mockClientKnex,
-  pluginMetadataCollectorMock,
   roleMetadataStorageMock,
 } from './mock-utils';
 import { clearAuditorMock } from './auditor-test-utils';
@@ -91,11 +90,9 @@ export function newConfig(
   });
 }
 
-export async function newAdapter(config: Config): Promise<Adapter> {
-  return await new CasbinDBAdapterFactory(
-    config,
-    mockClientKnex,
-  ).createAdapter();
+export async function newAdapter(_config: Config): Promise<Adapter> {
+  const testKnex = await createTestCasbinKnex();
+  return await CasbinKnexAdapter.newAdapter(testKnex);
 }
 
 export async function createEnforcer(
@@ -166,7 +163,6 @@ export async function newPermissionPolicy(
     enfDelegate,
     roleMock || roleMetadataStorageMock,
     mockClientKnex,
-    pluginMetadataCollectorMock as PluginPermissionMetadataCollector,
     mockUserInfoService,
     mockAuthService,
   );