+++ workspaces/announcements/plugins/announcements/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage-community/plugin-announcements
 
+## 2.13.0
+
+### Minor Changes
+
+- bae4489: Backstage version bump to v1.54.5
+
+### Patch Changes
+
+- Updated dependencies [bae4489]
+  - @backstage-community/plugin-announcements-common@0.24.0
+  - @backstage-community/plugin-announcements-react@0.29.0
+
+## 2.12.1
+
+### Patch Changes
+
+- fcd0f1a: Updated entity extension default filter to use the modern object form.
+
 ## 2.12.0
 
 ### Minor Changes